Refacto and change endpoint
