+++ www/front_src/src/Resources/Filter/api/index.ts
@@ -1,8 +1,5 @@
-import { JsonDecoder } from 'ts.data.json';
-
 import {
   buildListingEndpoint,
-  buildListingDecoder,
   getData,
   ListingModel,
   postData,
@@ -10,58 +7,12 @@ import {
 } from '@centreon/ui';
 
 import { baseEndpoint } from '../../api/endpoint';
-import { RawFilter, RawCriteria, CriteriaValue } from '../models';
-import { toRawFilter } from '../adapters';
+import { RawFilter } from '../models';
+import { toRawFilter } from './adapters';
 
-// const filterEndpoint = `${baseEndpoint}/users/filters/events-view`;
+const filterEndpoint = `${baseEndpoint}/users/filters/events-view`;
 
-const filterEndpoint = 'http://localhost:5000/mock/filters';
-
-const entityDecoder = JsonDecoder.object<RawFilter>(
-  {
-    id: JsonDecoder.number,
-    name: JsonDecoder.string,
-    criterias: JsonDecoder.array<RawCriteria>(
-      JsonDecoder.object<RawCriteria>(
-        {
-          name: JsonDecoder.string,
-          objectType: JsonDecoder.optional(JsonDecoder.string),
-          type: JsonDecoder.string,
-          value: JsonDecoder.oneOf<string | boolean | Array<CriteriaValue>>(
-            [
-              JsonDecoder.string,
-              JsonDecoder.boolean,
-              JsonDecoder.array<CriteriaValue>(
-                JsonDecoder.object<CriteriaValue>(
-                  {
-                    id: JsonDecoder.oneOf<number | string>(
-                      [JsonDecoder.number, JsonDecoder.string],
-                      'string | id ',
-                    ),
-                    name: JsonDecoder.string,
-                  },
-                  'FilterCriteriaValue',
-                ),
-                'FilterCriteriaValues',
-              ),
-            ],
-            'CriteriaIdValue',
-          ),
-        },
-        'FilterCriterias',
-        { objectType: 'object_type' },
-      ),
-      'FilterCriterias',
-    ),
-  },
-  'CustomFilter',
-);
-
-const listCustomFiltersDecoder = buildListingDecoder({
-  entityDecoder,
-  entityDecoderName: 'CustomFilter',
-  listingDecoderName: 'CustomFilters',
-});
+// const filterEndpoint = 'http://localhost:5000/mock/filters';
 
 const buildListCustomFiltersEndpoint = (params): string =>
   buildListingEndpoint({
@@ -69,11 +20,11 @@ const buildListCustomFiltersEndpoint = (params): string =>
     params,
   });
 
-const listCustomFilters = (cancelToken) => (
-  params,
-): Promise<ListingModel<RawFilter>> =>
+const listCustomFilters = (cancelToken) => (): Promise<
+  ListingModel<RawFilter>
+> =>
   getData<ListingModel<RawFilter>>(cancelToken)(
-    buildListCustomFiltersEndpoint(params),
+    buildListCustomFiltersEndpoint({ limit: 100, page: 1 }),
   );
 
 const createFilter = (cancelToken) => (params): Promise<number> => {
@@ -85,14 +36,13 @@ const createFilter = (cancelToken) => (params): Promise<number> => {
 
 const updateFilter = (cancelToken) => (params): Promise<void> =>
   patchData<Omit<RawFilter, 'id'>, void>(cancelToken)({
-    endpoint: filterEndpoint,
+    endpoint: `${filterEndpoint}/${params.id}`,
     data: toRawFilter(params),
   });
 
 export {
   listCustomFilters,
   buildListCustomFiltersEndpoint,
-  listCustomFiltersDecoder,
   createFilter,
   updateFilter,
 };
